+++ internal/query/client_test.go
@@ -960,7 +960,7 @@ func TestClient(t *testing.T) {
 					if want.a == 5 {
 						require.NoError(t, err)
 					}
-					if errors.Is(err, io.EOF) {
+					if xerrors.Is(err, io.EOF) {
 						require.Fail(t, "unexpected EOF")
 					}
 					var a uint64
@@ -1083,7 +1083,7 @@ func TestClient(t *testing.T) {
 					if want.a == 5 {
 						require.NoError(t, err)
 					}
-					if errors.Is(err, io.EOF) {
+					if xerrors.Is(err, io.EOF) {
 						require.Fail(t, "unexpected EOF")
 					}
 					var a uint64
@@ -1204,7 +1204,7 @@ func TestClient(t *testing.T) {
 					{5, "5"},
 				} {
 					row, err := rs.NextRow(ctx)
-					if errors.Is(err, io.EOF) {
+					if xerrors.Is(err, io.EOF) {
 						require.Fail(t, "unexpected EOF in RS0")
 					}
 					require.NoError(t, err)